+++ src/sapphire_config/metadata.py
@@ -1,3 +1,3 @@
 """Package metadata"""
 
-__version__ = "1.0.3"
+__version__ = "1.0.4"